refactor(api): extract data source, metamodel, knowledge and model routers
api.py drops from 2548 to 1180 lines across five routers grouped by concern
rather than by size. Each grouping is a claim about what belongs together:

- data sources stop at "we can read this system and know what is in it"
- the metamodel routes are all declarations the platform can explain, never
  inferences -- which is the basis of the generality ceiling (ADR-0005)
- knowledge and conversations share a module because they share one gate: an
  entry is not retrievable until confirmed and anchored, and citations come only
  from entries that passed it
- model config, AI drafting and the governance ledgers share the boundary that
  the model is an adviser, not an authority; the ledgers are what make that
  verifiable after the fact

Route count is unchanged at 135, verified after each extraction. Two tests
imported handlers from `ontology_platform.api` by name and moved with them.
